% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gglegends.R
\name{add_legend}
\alias{add_legend}
\title{Adds legend to plot}
\usage{
add_legend(gtab, legends, height = NA, width = NA, pos = "right")
}
\arguments{
\item{gtab}{a gtable}

\item{legends}{legend grobs}

\item{height}{height of column in which legend is added (only use if pos is "right")}

\item{width}{width of column in which legend is added (only use if pos is "bottom" or "top")}
}
\value{
gtable
}
\description{
Adds legend to plot
}
\seealso{
\code{\link{get_legend}}
}

