% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_ggplots.R
\name{align_plots_hor}
\alias{align_plots_hor}
\title{Aligns a set of plots horizontally.}
\usage{
align_plots_hor(..., widths = NA, plot = TRUE)
}
\arguments{
\item{...}{some ggplots (or gtables)}

\item{widths}{the width of each plot in output}

\item{plot}{print plot?}
}
\value{
If plot is True, then makes plot and returns gtable of plot invisibly.  Else, just returns gtable
}
\description{
Aligns a set of plots horizontally.
}
\seealso{
\code{\link{align_plots_vert}} and \code{\link{plot_custom_grid}}
}

