% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/align_ggplots.R
\name{align_plots_vert}
\alias{align_plots_vert}
\title{Aligns a set of plots vertically.}
\usage{
align_plots_vert(..., heights = NA, plot = T)
}
\arguments{
\item{...}{some ggplots (or gtables)}

\item{heights}{the height of each plot in output}
}
\value{
If plot is True, then makes plot and returns gtable of plot invisibly.  Else, just returns gtable
}
\description{
Aligns a set of plots vertically.
}
\seealso{
\code{\link{align_plots_hor}} and \code{\link{plot_custom_grid}}
}

