% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty.R
\name{draw_plot}
\alias{draw_plot}
\title{Draw a plot based on a gtable}
\usage{
draw_plot(x)
}
\arguments{
\item{x}{a gtable or other object that can be passed to grid.draw}
}
\value{
draws a plot
}
\description{
Draw a plot based on a gtable
}

