% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/align_ggplots.R
\name{force_widths_equal}
\alias{force_widths_equal}
\title{Force widths of sub-elements of plots to be equal}
\usage{
force_widths_equal(...)
}
\arguments{
\item{...}{some ggplots (or gtables)}
}
\value{
list of gtables
}
\description{
Force widths of sub-elements of plots to be equal
}
\seealso{
\code{\link{align_plots_vert}} and \code{\link{plot_custom_grid}}
}

