% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglegends.R
\name{get_legend}
\alias{get_legend}
\title{Gets legend from plot}
\usage{
get_legend(gplot, height = NA, width = NA)
}
\arguments{
\item{gplot}{a ggplot}

\item{height}{desired value of legend.key.height}

\item{width}{desired value of legend.key.width}
}
\value{
returns grob
}
\description{
Gets legend from plot
}
\seealso{
\code{\link{add_legend}}
}

