% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/align_ggplots.R
\name{insert.unit}
\alias{insert.unit}
\title{Function from gtable utils.r}
\usage{
insert.unit(x, values, after = length(x))
}
\arguments{
\item{x}{list of units}

\item{values}{units to add to x}

\item{after}{number of elements into x to add values}
}
\value{
List of units
}
\description{
Function from gtable utils.r
}

