% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gglegends.R
\docType{data}
\name{no_legend}
\alias{no_legend}
\title{Add to ggplot to remove legend}
\format{\preformatted{List of 1
 $ legend.position: chr "none"
 - attr(*, "class")= chr [1:2] "theme" "gg"
 - attr(*, "complete")= logi FALSE
 - attr(*, "validate")= logi TRUE
}}
\usage{
no_legend
}
\arguments{
\item{none}{}
}
\value{
theme
}
\description{
Add to ggplot to remove legend
}
\keyword{datasets}

