% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_ggplots.R
\name{plot_custom_grid}
\alias{plot_custom_grid}
\title{Make a custom grid of plots}
\usage{
plot_custom_grid(..., nrow = 1, ncol = 1, heights = NA, widths = NA,
  plot = TRUE)
}
\arguments{
\item{...}{some ggplots (or gtables)}

\item{nrow}{number of rows}

\item{ncol}{number of columns}

\item{heights}{the heights of each row (must correspond to nrow)}

\item{widths}{the width of each plot in output (must correspond to ncol)}

\item{plot}{print plot?}
}
\value{
If plot is True, then makes plot and returns gtable of plot invisibly.  Else, just returns gtable
}
\description{
Make a custom grid of plots
}
\seealso{
\code{\link{align_plots_vert}} and \code{\link{align_plots_hor}}
}

