% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty.R
\name{pub_theme}
\alias{pub_theme}
\title{Theme for scientific publication}
\usage{
pub_theme(base_size = 8, base_family = "Helvetica")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{font family}
}
\value{
ggplot2 theme
}
\description{
Theme for scientific publication
}

