% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplotfriend.R
\name{rplotfriend}
\alias{rplotfriend}
\title{rplotfriend}
\description{
Package with some plotting functions to  help make aesthetically pleasing plots.
}
\details{
Plotting functions use gggplot, gtable, and grid packages.  Includes functions for making heatmaps, aligning multiple distinct plots, formatting numbers, and manipulating legends.
}
\author{
Alicia Schep
}
\seealso{
\code{\link{align_plots_vert}}, \code{\link{align_plots_hor}}, \code{\link{plot_custom_grid}},
\code{\link{ggheatmap}}, \code{\link{pub_theme}}, \code{\link{pretty_scientific}}, \code{\link{get_legend}}, \code{\link{add_legend}}
}

