% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggheatmap.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Makes heatmap}
\usage{
ggheatmap(X, xlabel = NA, ylabel = NA, guide = "Value", name = NA,
  colors = c(scales::muted("blue"), "white", scales::muted("red")),
  limits = NA, fixed = T, guidebreaks = NA)
}
\arguments{
\item{X}{matrix with values to plot as heatmap}

\item{xlabel}{label for x axis}

\item{guide}{label for guide}

\item{name}{title}

\item{colors}{colors for scale}

\item{limits}{limits for y axis}

\item{fixed}{wether x and y axes should be on same scale}

\item{guidebreaks}{breaks for color scale}

\item{yabel}{label for y axis}
}
\value{
ggplot
}
\description{
Makes heatmap
}

