% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglegends.R
\docType{data}
\name{no_legend}
\alias{no_legend}
\title{Add to ggplot to remove legend}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 1.}
\usage{
no_legend
}
\description{
Add to ggplot to remove legend
}
\keyword{datasets}

