% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pretty.R
\name{pub_theme}
\alias{pub_theme}
\title{Theme for scientific publication}
\usage{
pub_theme(base_size = 8, base_family = "Helvetica")
}
\arguments{
\item{base_size}{}

\item{base_family}{}
}
\value{
ggplot2 theme
}
\description{
Theme for scientific publication
}

